@@ -1,17 +1,8 @@
-// This file can be replaced during build by using the `fileReplacements` array.
-// `ng build --prod` replaces `environment.ts` with `environment.prod.ts`.
-// The list of file replacements can be found in `angular.json`.
-
 export const environment = {
-  production: false,
-  apiEndpoint: "/api"
+    production: false,
+    apiEndpoint: '/api',
+    angular: 'Angular 18',
+    bootstrap: 'Bootstrap 5',
+    expressjs: 'Express.js 4.17.1',
+    mongoDb : 'MongoDB 7.0',
 };
-
-/*
- * For easier debugging in development mode, you can import the following file
- * to ignore zone related error stack frames such as `zone.run`, `zoneDelegate.invokeTask`.
- *
- * This import should be commented out in production mode because it will have a negative impact
- * on performance if an error is thrown.
- */
-// import 'zone.js/plugins/zone-error';  // Included with Angular CLI.
